Add a check for teams.
